backend/requirements.txt
flask
flask-socketio
pymongo
python-dotenv
google-generativeai
eventlet
